Move IStructureComponent file to better package
