+++ structures/structure-portcullis/src/main/java/nl/pim16aap2/animatedarchitecture/structures/portcullis/Portcullis.java
@@ -5,8 +5,8 @@ import lombok.ToString;
 import lombok.extern.flogger.Flogger;
 import nl.pim16aap2.animatedarchitecture.core.animation.AnimationRequestData;
 import nl.pim16aap2.animatedarchitecture.core.animation.IAnimationComponent;
+import nl.pim16aap2.animatedarchitecture.core.structures.IStructureComponent;
 import nl.pim16aap2.animatedarchitecture.core.structures.IStructureConst;
-import nl.pim16aap2.animatedarchitecture.core.structures.properties.IStructureComponent;
 import nl.pim16aap2.animatedarchitecture.core.structures.properties.Property;
 import nl.pim16aap2.animatedarchitecture.core.util.Cuboid;
 import nl.pim16aap2.animatedarchitecture.core.util.MovementDirection;
